AI: GraphSAGE inductive and transductive share the same inference process for dist
